ai_backend/requirements.txt
fastapi
torch
opencv-python-headless
uvicorn
numpy
flask
Pillow
pandas
requests
python-multipart
gunicorn